# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RUMIN API",
    "summary": "Financial intelligence and economic simulation platform.",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/providers": {
      "get": {
        "tags": [
          "providers and datasets"
        ],
        "summary": "List data providers",
        "description": "Who publishes the data RUMIN stores, with licensing, commercial-use and rate-limit notes taken from each provider's documentation.",
        "operationId": "list_providers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderRead"
                  },
                  "type": "array",
                  "title": "Response List Providers"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{provider_id}": {
      "get": {
        "tags": [
          "providers and datasets"
        ],
        "summary": "Get a data provider",
        "operationId": "get_provider",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{1,63}$",
              "examples": [
                "worldbank"
              ],
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRead"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "providers and datasets"
        ],
        "summary": "List datasets",
        "description": "Every dataset loaded into RUMIN — the illustrative curated network and provider data — with its licence, attribution and latest ingestion job.",
        "operationId": "list_datasets",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DatasetKind"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this kind.",
              "title": "Kind"
            },
            "description": "Only this kind."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetPage"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "providers and datasets"
        ],
        "summary": "Get a dataset",
        "operationId": "get_dataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9_-]{1,63}$",
              "examples": [
                "worldbank-wdi"
              ],
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetRead"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/economic-series": {
      "get": {
        "tags": [
          "economic series"
        ],
        "summary": "List economic series",
        "description": "Series RUMIN retrieves from providers, with coverage, the latest reported value and the outcome of the last retrieval. Historical data only — nothing is live.",
        "operationId": "list_series",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9_-]{1,63}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only items from this dataset.",
              "title": "Dataset Id"
            },
            "description": "Only items from this dataset."
          },
          {
            "name": "country_iso3",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Z]{3}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 3166-1 alpha-3 code.",
              "title": "Country Iso3"
            },
            "description": "ISO 3166-1 alpha-3 code."
          },
          {
            "name": "measure_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeasureType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measure Type"
            }
          },
          {
            "name": "frequency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Frequency"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Frequency"
            }
          },
          {
            "name": "has_data",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only series with (true) or without (false) values.",
              "title": "Has Data"
            },
            "description": "Only series with (true) or without (false) values."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search names and provider codes.",
              "title": "Q"
            },
            "description": "Search names and provider codes."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "name",
                "-name",
                "last_period",
                "-last_period",
                "provider_code"
              ],
              "type": "string",
              "description": "Order; '-' means descending.",
              "default": "name",
              "title": "Sort"
            },
            "description": "Order; '-' means descending."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesPage"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/economic-series/{series_id}": {
      "get": {
        "tags": [
          "economic series"
        ],
        "summary": "Get an economic series",
        "operationId": "get_series",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{2,95}$",
              "examples": [
                "wb-ind-fp-cpi-totl-zg"
              ],
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/economic-series/{series_id}/observations": {
      "get": {
        "tags": [
          "economic series"
        ],
        "summary": "List a series' observations",
        "description": "Values in period order, exactly as published. Periods the provider listed without a value are included with `value: null` (never filled in). Each value records when it was retrieved and the stored response it came from.",
        "operationId": "list_observations",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{2,95}$",
              "examples": [
                "wb-ind-fp-cpi-totl-zg"
              ],
              "title": "Series Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First date to include (YYYY-MM-DD).",
              "title": "Start"
            },
            "description": "First date to include (YYYY-MM-DD)."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last date to include (YYYY-MM-DD).",
              "title": "End"
            },
            "description": "Last date to include (YYYY-MM-DD)."
          },
          {
            "name": "include_missing",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include periods without a value.",
              "default": true,
              "title": "Include Missing"
            },
            "description": "Include periods without a value."
          },
          {
            "name": "include_revisions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return superseded revisions (the full history of what was received).",
              "default": false,
              "title": "Include Revisions"
            },
            "description": "Also return superseded revisions (the full history of what was received)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationPage"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/instruments": {
      "get": {
        "tags": [
          "market data"
        ],
        "summary": "List instruments",
        "description": "Listed instruments declared by imported price files. RUMIN ships none.",
        "operationId": "list_instruments",
        "parameters": [
          {
            "name": "exchange_mic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Z0-9]{4}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "E.g. XNSE or XBOM.",
              "title": "Exchange Mic"
            },
            "description": "E.g. XNSE or XBOM."
          },
          {
            "name": "instrument_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InstrumentType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instrument Type"
            }
          },
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9_-]{1,63}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only items from this dataset.",
              "title": "Dataset Id"
            },
            "description": "Only items from this dataset."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentPage"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/instruments/{instrument_id}": {
      "get": {
        "tags": [
          "market data"
        ],
        "summary": "Get an instrument",
        "operationId": "get_instrument",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{2,95}$",
              "examples": [
                "xnse-reliance"
              ],
              "title": "Instrument Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentDetail"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetKind": {
        "type": "string",
        "enum": [
          "curated",
          "provider"
        ],
        "title": "DatasetKind",
        "description": "Where a dataset's records came from.\n\n* ``curated`` — reference data written for RUMIN and loaded from a file in the\n  repository (the Phase 1 illustrative network).\n* ``provider`` — data retrieved from an external provider (an API or a file the user\n  is licensed to use), with the provider's licence and attribution."
      },
      "DatasetPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total number of items matching the query."
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "DatasetPage",
        "description": "A page of datasets."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Every non-2xx response from the API uses this envelope."
      },
...
      "Frequency": {
        "type": "string",
        "enum": [
          "daily",
          "weekly",
          "monthly",
          "quarterly",
          "annual",
          "irregular"
        ],
        "title": "Frequency"
      },
...
      "InstrumentType": {
        "type": "string",
        "enum": [
          "equity",
          "etf",
          "index"
        ],
        "title": "InstrumentType"
      },
...
    }
  },
...
}